Move data structures to ical.go
These will be shared between the parser and the formatter.

While at it, add a package description.
